Add files via upload
Red, Yellow, Green LEDs ON for every 5 seconds. Can be observed in serial monitor.
When RFID card is detected by EM-18 Reader module, Green LED glows deviating from the original sequence.
If the RFID card number (12-digit) is not stored initially then there is no change in the signal (but displays the RFID card number in the serial monitor) and original sequence goes on.